Orgenize interfaces and variables
