+++ src/app/interfaces/information-screen/informationScreen.interface.ts
@@ -1,0 +1,12 @@
+export enum InformationScreenType {
+  Error = "error",
+  Loading = "loading",
+  NoConnection = "no-connection",
+  Unknown = "unknown",
+}
+
+export interface IInformationScreenData {
+  type?: InformationScreenType;
+  error?: string;
+  isWebApp?: boolean;
+}
